XRE-11535 : Implemented XUPNP Update notification for FOG
Reason for change: Fog waits indefinity for STMgr to come up first n report TSB is supported or NOT.. Only then Fog inits.. Buts XUPNP starts in parallel and it checks the Fog status only once when launching. This leads to wrongly advertise that Fog is not available as it was waiting for STMGR when XUPNP queries. Added Event Handler to listen n post update.
Test Procedure: As per XRE-11535
Risks: Low
Signed-off-by: Karunakaran A <Karunakaran_Amirthalingam@cable.comcast.com>

Change-Id: Ice9caa5de615f64011ed98c21e6934eb886e8456