RDK-24171 : UPnP DeviceProtection Service and Authentication
Reason for change: Check in device protection changes for device
and control point. Adding the CA certificate and key files for
all devices and the script to generate the csr and keys
based on the CA certificates
Test Procedure: Test with device protection enabled/disabled,
check if the control point connection to the device.
Risks: None
Signed-off-by: pmathe686 <philip_mathew@comcast.com>

RDKB-24404 : UPnP DeviceProtection Service and Authentication

Reason for change: Copy certificates and key files for device protection.
Test Procedure: Test with device protection enabled/disabled,
check if the control point connection to the device is successful.
Risks: None
Signed-off-by: pmathe686 <philip_mathew@comcast.com>
Change-Id: Iedb5fb0b0cfb67cdf874420051f4af315cfa4365

RDKB-25101 : UPnP Device Protection and Refactor not appearing

Reason for change: In broadband devices, the script generating
the certificates and
keys was using ntp sync status for ensuring that the time is in sync
before generating the certificates, the ntp status was not working
correctly. Modified to use wan status instead of ntp status.
Verified on broadband devices
Test Procedure: Check the date of the certificate (ice*_t and ice*_y)
 and key generated in /tmp folder of the device after device power cycleTest
with device protection enabled/disabled,
check control point connection to the device is successful.
Risks: None
Signed-off-by: pmathe686 <philip_mathew@comcast.com>

RDKB-25101: UPnP DP Refactor not enabling for xdiscovery on bootup

Reason for change: If the device is in ethwan mode the wan_status flag used
for checking the network timesync is not working since
in this case this flag will not get updated.
Hence increased the delay to accomodate these device.
Test Procedure: check for upnp rfc whith rfc enabled.
Risks: Low

Signed-off-by: Kailash Baskar <kailash_baskar@comcast.com>

RDKB-24530 RDK-24171 : UPnP DeviceProtection Service and

Reason for change:To initiate dpcg.sh script on video devices
and to correct RDKB-24530 issue on Broadband devices.
Test Procedure: Test with device protection enabled/disabled,
check if the control point connection to the device.
Risks: None
Signed-off-by: kbaska201 <kailash_baskar@comcast.com>

RDKB-25101: UPnP DP  Refactor not enabling for xdiscovery on bootup

Reason for change:Xdiscovery service with DP not working when rfc is turned on.
Test Procedure: Reboot the box with RFC enabled, ensure xdiscovery
is working with DP.
Risks: Low
Signed-off-by: kbaska201 <kailash_baskar@comcast.com>

Signed-off-by: Thomson Naveen Silvera <thomson_silvera@comcast.com>
Change-Id: Ic0ebf0ca66fc5d76518a915fd43fe77740b0859b