RDKB-40548:Observed error message "dnsconfig file name not found in
             config" in xdevice.log
Reason for change: Due to CLIENT_XCAL_SERVER defined in broadband devices dnsConf value is not populated in the structure.So skipping the retrieval part as it is not used anywhere else.
Test Procedure: No "file name not found" should be logged in xdevice.log in both normal upnp and device protection mode.
Risks: Low
Signed-off-by:manish_lepuru@comcast.com

Change-Id: I447fc4890362af633c61722ca89d462f0026316d